Cache stuff

+++ code/controllers/subsystem/liquids.dm
@@ -1,6 +1,6 @@
 SUBSYSTEM_DEF(liquids)
 	name = "Liquid Turfs"
-	wait = 0.5 SECONDS
+	wait = 1 SECONDS
 	flags = SS_BACKGROUND
 	runlevels = RUNLEVEL_GAME | RUNLEVEL_POSTGAME
 
@@ -29,7 +29,7 @@ SUBSYSTEM_DEF(liquids)
 			currentrun_active_turfs -= T //work off of index later
 		if(!currentrun_active_turfs.len)
 			run_type = SSLIQUIDS_RUN_TYPE_GROUPS
-	else if (run_type == SSLIQUIDS_RUN_TYPE_GROUPS)
+	if (run_type == SSLIQUIDS_RUN_TYPE_GROUPS)
 		for(var/g in active_groups)
 			var/datum/liquid_group/LG = g
 			if(LG.dirty)
@@ -69,6 +69,7 @@ SUBSYSTEM_DEF(liquids)
 		var/volume_change = R.volume * fraction
 		liquids.reagents.remove_reagent(R.type, volume_change)
 		T.add_liquid(R.type, volume_change, TRUE)
+	liquids.has_cached_share = FALSE
 
 /turf/proc/liquid_update_turf()
 	//Check atmos adjacency to cut off any disconnected groups
@@ -91,10 +92,10 @@ SUBSYSTEM_DEF(liquids)
 	liquids.reagents.add_reagent(reagent, amount, no_react = no_react)
 	liquids.calculate_height()
 	liquids.set_reagent_color_for_liquid()
+	liquids.has_cached_share = FALSE
 	SSliquids.add_active_turf(src)
 	if(lgroup)
 		lgroup.dirty = TRUE
-
 
 /obj/effect/abstract/liquid_turf
 	name = "liquid"
@@ -109,6 +110,7 @@ SUBSYSTEM_DEF(liquids)
 	var/only_big_diffs = 1
 	var/turf/my_turf
 	var/liquid_state = LIQUID_STATE_PUDDLE
+	var/has_cached_share = FALSE
 
 /obj/effect/abstract/liquid_turf/update_overlays()
 	. = ..()
@@ -189,12 +191,15 @@ SUBSYSTEM_DEF(liquids)
 	var/amount_of_active_turfs = 0
 	var/decay_counter = 0
 	var/expected_turf_height = 0
+	var/list/last_cached_fraction_share
 
 /datum/liquid_group/proc/add_to_group(turf/T)
 	members[T] = TRUE
 	T.lgroup = src
 	if(SSliquids.active_turfs[T])
 		amount_of_active_turfs++
+	if(T.liquids)
+		T.liquids.has_cached_share = FALSE
 
 /datum/liquid_group/proc/remove_from_group(turf/T)
 	members -= T
@@ -241,7 +246,9 @@ SUBSYSTEM_DEF(liquids)
 	current_adjacent[T] = TRUE
 	recursive_adjacent[T] = TRUE
 	var/getting_new_turfs = TRUE
-	while(getting_new_turfs)
+	var/indef_loop_safety = 0
+	while(getting_new_turfs && indef_loop_safety < LIQUID_RECURSIVE_LOOP_SAFETY)
+		indef_loop_safety++
 		getting_new_turfs = FALSE
 		var/list/new_adjacent = list()
 		for(var/t in current_adjacent)
@@ -265,46 +272,88 @@ SUBSYSTEM_DEF(liquids)
 /datum/liquid_group/proc/share(use_liquids_color = FALSE)
 	var/datum/reagents/tempr = new(1000000)
 	var/any_share = FALSE
+	var/cached_shares = 0
+	var/list/cached_add = list()
+	//log_game("BEGIN SHARE:.")
 	for(var/t in members)
 		//var/share_type = members[t]
 		var/turf/T = t
 		if(T.liquids/* && share_type == LIQUID_MUTUAL_SHARE*/)
 			any_share = TRUE
-			T.liquids.reagents.trans_to(tempr, T.liquids.reagents.total_volume, no_react = TRUE)
+			
+			if(T.liquids.has_cached_share && last_cached_fraction_share)
+				cached_shares++
+				continue
+			
+			//T.liquids.reagents.trans_to(tempr, T.liquids.reagents.total_volume, no_react = TRUE)
+			//log_game("LIQUIDS FROM MEMBER:.")
+			for(var/r in T.liquids.reagents.reagent_list)
+				var/datum/reagent/R = r
+				if(!cached_add[R.type])
+					cached_add[R.type] = 0
+				//log_game("[R.type] += [R.volume].")
+				cached_add[R.type] += R.volume
 	if(!any_share)
 		return
 	decay_counter = 0
+
 	
-	var/fraction = 1 / members.len
+	if(cached_shares)
+		for(var/reagent_type in last_cached_fraction_share)
+			if(!cached_add[reagent_type])
+				cached_add[reagent_type] = 0
+			cached_add[reagent_type] += last_cached_fraction_share[reagent_type] * cached_shares
+	
+	/*
+	log_game("TOTAL REAGENTS:.")
+	for(var/reagent_type in cached_add)
+		log_game("[reagent_type] = [cached_add[reagent_type]].")
+	log_game("FRACTIONED REAGENTS (MEMBERS: [members.len]):.")*/
+	for(var/reagent_type in cached_add)
+		cached_add[reagent_type] = cached_add[reagent_type] / members.len
+		//log_game("[reagent_type] = [cached_add[reagent_type]].")
+	last_cached_fraction_share = cached_add
+	tempr.add_reagent_list(cached_add, no_react = TRUE)
 	var/mixed_color = use_liquids_color ? mix_color_from_reagents(tempr.reagent_list) : color
-	var/height = CEILING(tempr.total_volume/members.len/LIQUID_HEIGHT_DIVISOR, 1)
-	
+	var/height = CEILING(tempr.total_volume/LIQUID_HEIGHT_DIVISOR, 1)
+
 	//message_admins("shared")
 	for(var/t in members)
 		var/turf/T = t
 		if(!T.liquids)
 			T.liquids = new(T)
 		var/obj/effect/abstract/liquid_turf/cached_liquids = T.liquids
-		tempr.copy_to(cached_liquids.reagents, tempr.total_volume, fraction, no_react = TRUE)
-		
+		T.liquids.create_reagents(1000) //Computing nice way to clean the old ones
+		tempr.copy_to(cached_liquids.reagents, tempr.total_volume, no_react = TRUE)
+		T.liquids.has_cached_share = TRUE
+
 		cached_liquids.set_height(height)
 		cached_liquids.color = mixed_color
 	qdel(tempr)
 
 /datum/liquid_group/proc/process_cell(turf/T)
+	if(T.liquids.height <= 1)
+		return FALSE
 	for(var/tur in T.GetAtmosAdjacentTurfs())
 		var/turf/T2 = tur
 		if(T.z != T2.z)
 			var/turf/Z_turf_below = SSmapping.get_turf_below(T)
-			var/turf/Z_turf_above = SSmapping.get_turf_above(T)
 			if(T2 == Z_turf_below)
 				if(!(T2.liquids && T2.liquids.height >= LIQUID_HEIGHT_CONSIDER_FULL_TILE))
 					T.liquid_fraction_share(T2, 1)
 					qdel(T.liquids, TRUE)
 					. = TRUE
-					continue
-			else if(T2 == Z_turf_above)
-				continue
+			continue
+		//CHECK DIFFERENT TURF HEIGHT THING
+		if(T.liquid_height != T2.liquid_height)
+			var/my_liquid_height = T.liquid_height + T.liquids.height
+			var/target_liquid_height = T2.liquid_height + (T2.liquids ? T2.liquids.height : 0)
+			if(my_liquid_height > target_liquid_height+1)
+				var/height_diff = abs((target_liquid_height / my_liquid_height)-1)
+				T.liquid_fraction_share(T2, height_diff)
+				. = TRUE
+			continue
+		//END OF TURF HEIGHT
 		if(!T.can_share_liquids_with(T2))
 			continue
 		if(!T2.lgroup)
@@ -330,7 +379,17 @@ SUBSYSTEM_DEF(liquids)
 		return FALSE
 	*/
 	var/my_liquid_height = liquids ? liquids.height : 0
+	if(my_liquid_height < 1)
+		return FALSE
 	var/target_height = T.liquids ? T.liquids.height : 0
+
+	//Varied heights handling:
+	if(liquid_height != T.liquid_height)
+		if(my_liquid_height+liquid_height < target_height + T.liquid_height + 1)
+			return FALSE
+		else
+			return TRUE
+
 	var/difference = abs(target_height - my_liquid_height)
 	//The: sand effect or "piling" Very good for performance
 	if(difference > 1) //SHOULD BE >= 1 or > 1? '>= 1' can lead into a lot of unnessecary processes, while ' > 1' will lead to a "piling" phenomena
@@ -342,7 +401,7 @@ SUBSYSTEM_DEF(liquids)
 		//Check if anything adjacent has liquids that we can share with
 		for(var/tur in GetAtmosAdjacentTurfs())
 			var/turf/T2 = tur
-			if(T2.liquids && T2.can_share_liquids_with(src))
+			if(T2.liquids && T2.can_share_liquids_with(src) && liquid_height == T2.liquid_height)
 				if(!T2.lgroup)
 					var/datum/liquid_group/new_group = new(liquid_height)
 					new_group.add_to_group(T2)
